* Changed double to float

+++ src/DungeonGame/AbstractEntity.java
@@ -8,10 +8,10 @@ import java.awt.*;
 
 public abstract class AbstractEntity implements Entity {
 
-    protected double x, y, width, height;
+    protected float x, y, width, height;
     protected Rectangle hitbox = new Rectangle();
 
-    public AbstractEntity(double x, double y, double width, double height) {
+    public AbstractEntity(float x, float y, float width, float height) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -19,48 +19,48 @@ public abstract class AbstractEntity implements Entity {
     }
 
     @Override
-    public void setLocation(double x, double y) {
+    public void setLocation(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
     @Override
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
     @Override
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
     @Override
-    public void setWidth(double width) {
+    public void setWidth(float width) {
         this.width = width;
     }
 
     @Override
-    public void setHeight(double height) {
+    public void setHeight(float height) {
         this.height = height;
     }
 
     @Override
-    public double getX() {
+    public float getX() {
         return x;
     }
 
     @Override
-    public double getY() {
+    public float getY() {
         return y;
     }
 
     @Override
-    public double getHeight() {
+    public float getHeight() {
         return height;
     }
 
     @Override
-    public double getWidth() {
+    public float getWidth() {
         return width;
     }
